+++ AI/DeepLearning/NegativeSampling.Test/NegativeSamplerShould.cs
@@ -1,5 +1,5 @@
 using NeuralNetwork;
-using NeuralNetwork.Data;
+using NeuralNetwork.Models;
 using System;
 using System.Collections.Generic;
 using Xunit;
