+++ AI/Models/NeuralNetwork/Models/Weight.cs
@@ -1,5 +1,8 @@
-﻿namespace NeuralNetwork.Data
+﻿using System;
+
+namespace NeuralNetwork.Models
 {
+    [Serializable]
     public class Weight
     {
         public Weight(double value)
