+++ AI/Tests/AI.Tests/Word2Vec/BasicNovelWithoutSoftmax.cs
@@ -3,7 +3,6 @@
     using System;
     using System.IO;
     using global::Word2Vec;
-    using Xunit;
 
     public class BasicNovelWithoutSoftmax
     {
